ggml : fix vector-index PR2a review issues
Preserve existing error-code values, keep quantized SIMD ranking in double precision, validate IVF state for empty searches, and align persistence/package error handling with the documented contracts.

+++ ggml/src/ggml-vector-index-avx2.h
@@ -4,7 +4,7 @@
 
 namespace ggml_vec_index_detail {
 
-float dot_q8_avx2(const float * query, const int8_t * codes, float scale, int dim);
-float dot_q4_avx2(const float * query, const uint8_t * codes, float scale, int dim);
+double dot_q8_avx2(const float * query, const int8_t * codes, float scale, int dim);
+double dot_q4_avx2(const float * query, const uint8_t * codes, float scale, int dim);
 
 } // namespace ggml_vec_index_detail
